PyCharm/poetry/pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["feus <barca105@nate.com>"]


[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.36"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
graphene-sqlalchemy = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
